Fixed warnings when compiling for 64-bit Windows.
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS/branches/WIN64@6237 212acab6-be3b-0410-9dea-997c60f758d6

+++ Contrib/Library/RegTool/RegTool.c
@@ -282,7 +282,7 @@ void RenameViaWininit(const TCHAR* prevName, const TCHAR* newName)
   LPCSTR szRenameSec = "[Rename]\r\n"; // rename section marker
   HANDLE hfile;
   DWORD dwFileSize;
-  DWORD dwBytes;
+  SIZE_T dwBytes;
   DWORD dwRenameLinePos;
   char *pszWinInit;   // Contains the file contents of wininit.ini
